refactor: use usize for coordinates and document edge cases

+++ src/editor.rs
@@ -64,8 +64,11 @@ impl Editor {
 
     fn draw_welcome_message() -> Result<()> {
         let welcome_message = format!("{NAME} editor -- version {VERSION}");
-        let width = Terminal::size()?.width as usize;
+        let width = Terminal::size()?.width;
         let len = welcome_message.len();
+
+        // We allow this since we don't care if our welcome message is put _exactly_ in the middle.
+        // It's allowed to be a bit to the left or right.
         let padding = width.saturating_sub(len) / 2;
         let space = " ".repeat(padding.saturating_sub(1));
         let mut welcome_message = format!("~{space}{welcome_message}");
@@ -81,6 +84,9 @@ impl Editor {
         let Size { height, .. } = Terminal::size()?;
         for current_row in 0..height {
             Terminal::clear_line()?;
+
+            // We allow this since we don't care if our welcome message is put _exactly_ in the middle.
+            // It's allowed to be a bit up or down
             if current_row == height / 3 {
                 Self::draw_welcome_message()?;
             } else {
